requirements.txt
# Telegram
telethon>=1.36.0

# Data & Models
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy>=2.0
alembic>=1.13

# Market Data
ccxt>=4.0
httpx>=0.27

# Analysis
pandas>=2.0
numpy>=1.26

# Dashboard
streamlit>=1.30

# Config
pyyaml>=6.0

# Scheduling
apscheduler>=3.10

# Dev & Test
pytest>=8.0
pytest-asyncio>=0.23
